Move admin panel files
