+++ src/features/adminPanel/subcomponents/assessmentConfigPanel/BooleanCell.tsx
