+++ src/features/adminPanel/subcomponents/assessmentConfigPanel/DeleteRowCell.tsx
