+++ src/features/adminPanel/subcomponents/assessmentConfigPanel/NumericCell.tsx
